Ok. Still no idea what's going on with quantisation, but I understand the layout, I think.
I'm pretty sure that colors are compressed independently of each other! So I should theoretically be able to unroll the loops and dramatically simplify some of the C++ code. I think that's a good next step.

This will probably be slightly easier using `nom` because there are bitwise parsers. Woohoo!